+++ app/src/main/java/com/vultisig/wallet/ui/models/send/SendFormViewModel.kt
@@ -14,30 +14,21 @@ import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
 import com.vultisig.wallet.R.string
 import com.vultisig.wallet.data.blockchain.FeeServiceComposite
-import com.vultisig.wallet.data.blockchain.model.StakingDetails.Companion.generateId
-import com.vultisig.wallet.data.blockchain.model.Transfer
-import com.vultisig.wallet.data.blockchain.model.VaultData
-import com.vultisig.wallet.data.blockchain.thorchain.RujiStakingService.Companion.RUJI_REWARDS_COIN
 import com.vultisig.wallet.data.blockchain.tron.GetTronFrozenBalancesUseCase
 import com.vultisig.wallet.data.blockchain.tron.TronFrozenBalanceState
 import com.vultisig.wallet.data.blockchain.tron.TronResourceType
-import com.vultisig.wallet.data.blockchain.tron.TronStakingOperation
-import com.vultisig.wallet.data.blockchain.tron.tronStakingMemo
 import com.vultisig.wallet.data.models.Account
 import com.vultisig.wallet.data.models.Address
 import com.vultisig.wallet.data.models.AddressBookEntry
 import com.vultisig.wallet.data.models.Chain
 import com.vultisig.wallet.data.models.ChainId
 import com.vultisig.wallet.data.models.Coin
-import com.vultisig.wallet.data.models.Coins
-import com.vultisig.wallet.data.models.GasFeeParams
 import com.vultisig.wallet.data.models.ImageModel
 import com.vultisig.wallet.data.models.TokenId
 import com.vultisig.wallet.data.models.TokenStandard
 import com.vultisig.wallet.data.models.TokenValue
 import com.vultisig.wallet.data.models.Vault
 import com.vultisig.wallet.data.models.VaultId
-import com.vultisig.wallet.data.models.getPubKeyByChain
 import com.vultisig.wallet.data.models.payload.BlockChainSpecific
 import com.vultisig.wallet.data.repositories.AccountsRepository
 import com.vultisig.wallet.data.repositories.AddressParserRepository
@@ -57,7 +48,6 @@ import com.vultisig.wallet.data.usecases.GasFeeToEstimatedFeeUseCase
 import com.vultisig.wallet.data.usecases.GetAvailableTokenBalanceUseCase
 import com.vultisig.wallet.data.usecases.RequestAddressBookEntryUseCase
 import com.vultisig.wallet.data.usecases.RequestQrScanUseCase
-import com.vultisig.wallet.data.utils.safeLaunch
 import com.vultisig.wallet.ui.models.mappers.AccountToTokenBalanceUiModelMapper
 import com.vultisig.wallet.ui.models.mappers.TokenValueToStringWithUnitMapper
 import com.vultisig.wallet.ui.models.send.AmountFraction.F100
@@ -82,38 +72,27 @@ import com.vultisig.wallet.ui.screens.select.AssetSelected
 import com.vultisig.wallet.ui.screens.v2.defi.model.DeFiNavActions
 import com.vultisig.wallet.ui.screens.v2.defi.model.parseDepositType
 import com.vultisig.wallet.ui.utils.UiText
-import com.vultisig.wallet.ui.utils.asAddressInput
 import com.vultisig.wallet.ui.utils.asUiText
-import com.vultisig.wallet.ui.utils.textAsFlow
 import dagger.hilt.android.lifecycle.HiltViewModel
-import java.math.BigDecimal
 import java.math.BigInteger
 import javax.inject.Inject
 import kotlin.uuid.Uuid
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.Job
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.debounce
-import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.mapNotNull
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import timber.log.Timber
 import wallet.core.jni.proto.Bitcoin
 
@@ -272,8 +251,6 @@ constructor(
 
     private var mscaAddress: String? = null
 
-    private val recalculateGasFee = MutableStateFlow(0L)
-
     private val selectedToken = MutableStateFlow<Coin?>(null)
 
     private val selectedTokenValue: Coin?
@@ -287,9 +264,6 @@ constructor(
             val accounts = accounts.value
             return accounts.find { it.token.id.equals(selectedTokenValue?.id, true) }
         }
-
-    private var preSelectTokenJob: Job? = null
-    private var loadAccountsJob: Job? = null
 
     private val appCurrency =
         appCurrencyRepository.currency.stateIn(
@@ -311,10 +285,40 @@ constructor(
 
     private val tronFrozenBalances =
         MutableStateFlow<TronFrozenBalanceState>(TronFrozenBalanceState.Loading)
-    private var loadTronFrozenBalancesJob: Job? = null
 
-    private fun isTronStakingType(): Boolean =
-        defiType == DeFiNavActions.FREEZE_TRX || defiType == DeFiNavActions.UNFREEZE_TRX
+    private val tokenPreselectionService =
+        TokenPreselectionService(
+            scope = viewModelScope,
+            accounts = accounts,
+            defiTypeProvider = { defiType },
+            selectedTokenProvider = { selectedTokenValue },
+            onTokenSelected = ::selectToken,
+        )
+
+    private val accountsLoader =
+        AccountsLoader(
+            scope = viewModelScope,
+            accounts = accounts,
+            accountsRepository = accountsRepository,
+            stakingDetailsRepository = stakingDetailsRepository,
+            defiTypeProvider = { defiType },
+            mscaAddressProvider = { mscaAddress },
+        )
+
+    private val tronStakingService =
+        TronStakingService(
+            scope = viewModelScope,
+            uiState = uiState,
+            tronFrozenBalances = tronFrozenBalances,
+            tokenAmountFieldState = tokenAmountFieldState,
+            fiatAmountFieldState = fiatAmountFieldState,
+            memoFieldState = memoFieldState,
+            defiTypeProvider = { defiType },
+            vaultProvider = { vault },
+            vaultIdProvider = { vaultId },
+            vaultRepository = vaultRepository,
+            getTronFrozenBalances = getTronFrozenBalances,
+        )
 
     private val addressManager =
         AddressManager(
@@ -350,6 +354,36 @@ constructor(
 
     private val bitcoinPlanService = BitcoinPlanService(vaultRepository)
 
+    private val gasFeeOrchestrator =
+        GasFeeOrchestrator(
+            scope = viewModelScope,
+            uiState = uiState,
+            selectedToken = selectedToken,
+            accounts = accounts,
+            gasFee = gasFee,
+            gasSettings = gasSettings,
+            specific = specific,
+            planFee = planFee,
+            planBtc = planBtc,
+            addressFieldState = addressFieldState,
+            tokenAmountFieldState = tokenAmountFieldState,
+            memoFieldState = memoFieldState,
+            vaultProvider = { vault },
+            vaultIdProvider = { vaultId },
+            accountProvider = { selectedAccount },
+            resolvedDstAddressProvider = { addressManager.resolvedDstAddress.value },
+            isMaxAmountFlow = amountManager.isMaxAmount,
+            feeServiceComposite = feeServiceComposite,
+            tokenRepository = tokenRepository,
+            addressParserRepository = addressParserRepository,
+            chainAccountAddressRepository = chainAccountAddressRepository,
+            blockChainSpecificRepository = blockChainSpecificRepository,
+            advanceGasUiRepository = advanceGasUiRepository,
+            gasFeeToEstimatedFee = gasFeeToEstimatedFee,
+            bitcoinPlanService = bitcoinPlanService,
+            mapTokenValueToString = mapTokenValueToString,
+        )
+
     private val amountFractionManager =
         AmountFractionManager(
             scope = viewModelScope,
@@ -363,7 +397,7 @@ constructor(
             defiTypeProvider = { defiType },
             vaultProvider = { vault },
             accountProvider = { selectedAccount },
-            currentTronFrozenBalanceProvider = ::currentTronFrozenBalance,
+            currentTronFrozenBalanceProvider = tronStakingService::currentFrozenBalance,
             getAvailableTokenBalance = getAvailableTokenBalance,
             feeServiceComposite = feeServiceComposite,
             tokenRepository = tokenRepository,
@@ -396,7 +430,7 @@ constructor(
             vaultIdProvider = { vaultId },
             selectedAccountProvider = { selectedAccount },
             defiTypeProvider = { defiType },
-            currentTronFrozenBalanceProvider = ::currentTronFrozenBalance,
+            currentTronFrozenBalanceProvider = tronStakingService::currentFrozenBalance,
             navigator = navigator,
             expandSection = ::expandSection,
             emitFocusField = { field -> _focusFieldChannel.trySend(field) },
@@ -549,15 +583,10 @@ constructor(
         )
         loadSelectedCurrency()
         collectSelectedAccount()
-        calculateGasFees()
-        calculateGasTokenBalance()
-        collectEstimatedFee()
-        collectPlanFee()
-        calculateSpecific()
+        gasFeeOrchestrator.start()
         collectAdvanceGasUi()
         loadVaultName()
         loadGasSettings()
-        collectMaxAmount()
         addressManager.start()
         amountManager.start()
         observeManagers()
@@ -612,7 +641,7 @@ constructor(
 
         if (this.vaultId != vaultId) {
             this.vaultId = vaultId
-            loadAccounts(vaultId)
+            accountsLoader.load(vaultId)
             loadVaultName()
             initFormType()
         }
@@ -624,7 +653,7 @@ constructor(
                 preSelectedTokenId = preSelectedTokenId,
             )
         } else {
-            preSelectToken(
+            tokenPreselectionService.preSelect(
                 preSelectedChainIds = listOf(preSelectedChainId),
                 preSelectedTokenId = preSelectedTokenId,
             )
@@ -650,7 +679,8 @@ constructor(
     private fun initFormType() {
         val autoCompound =
             defiType == DeFiNavActions.STAKE_STCY || defiType == DeFiNavActions.UNSTAKE_STCY
-        val initialResourceType = if (isTronStakingType()) TronResourceType.BANDWIDTH else null
+        val initialResourceType =
+            if (tronStakingService.isStakingType()) TronResourceType.BANDWIDTH else null
         uiState.update {
             it.copy(
                 defiType = this.defiType,
@@ -658,97 +688,10 @@ constructor(
                 tronResourceType = initialResourceType,
             )
         }
-        if (isTronStakingType()) {
-            applyTronStakingMemo(TronResourceType.BANDWIDTH)
-            if (defiType == DeFiNavActions.UNFREEZE_TRX) {
-                loadTronFrozenBalances()
-            }
-        }
+        tronStakingService.initIfStakingType()
     }
 
-    fun setTronResourceType(type: TronResourceType) {
-        // Ignore toggles while a send is in flight so the captured memo/resource type
-        // used by send() can't diverge from the one visible in the tab.
-        if (uiState.value.isLoading) return
-        if (uiState.value.tronResourceType == type) return
-        uiState.update {
-            it.copy(
-                tronResourceType = type,
-                tronBalanceAvailableOverride = null,
-                selectedAmountFraction = null,
-            )
-        }
-        applyTronStakingMemo(type)
-        tokenAmountFieldState.clearText()
-        fiatAmountFieldState.clearText()
-        if (defiType == DeFiNavActions.UNFREEZE_TRX) {
-            updateTronFrozenBalanceDisplay(type)
-        }
-    }
-
-    private fun applyTronStakingMemo(type: TronResourceType) {
-        val op =
-            when (defiType) {
-                DeFiNavActions.FREEZE_TRX -> TronStakingOperation.FREEZE
-                DeFiNavActions.UNFREEZE_TRX -> TronStakingOperation.UNFREEZE
-                else -> error("applyTronStakingMemo called outside Tron staking")
-            }
-        memoFieldState.setTextAndPlaceCursorAtEnd(tronStakingMemo(op, type))
-    }
-
-    private fun loadTronFrozenBalances() {
-        loadTronFrozenBalancesJob?.cancel()
-        loadTronFrozenBalancesJob =
-            viewModelScope.safeLaunch(
-                onError = { e ->
-                    Timber.e(e, "Failed to load Tron frozen balances")
-                    setTronFrozenBalanceState(TronFrozenBalanceState.Error)
-                }
-            ) {
-                uiState.update { it.copy(isTronFrozenBalancesLoading = true) }
-                try {
-                    val vault = vault ?: vaultRepository.get(vaultId ?: return@safeLaunch)
-                    val trxCoin =
-                        vault?.coins?.firstOrNull { it.chain == Chain.Tron && it.isNativeToken }
-                    if (trxCoin == null) {
-                        setTronFrozenBalanceState(TronFrozenBalanceState.Error)
-                        return@safeLaunch
-                    }
-                    val balances = getTronFrozenBalances(trxCoin.address)
-                    setTronFrozenBalanceState(TronFrozenBalanceState.Loaded(balances))
-                    updateTronFrozenBalanceDisplay(
-                        uiState.value.tronResourceType ?: TronResourceType.BANDWIDTH
-                    )
-                } finally {
-                    uiState.update { it.copy(isTronFrozenBalancesLoading = false) }
-                }
-            }
-    }
-
-    private fun setTronFrozenBalanceState(state: TronFrozenBalanceState) {
-        tronFrozenBalances.value = state
-        uiState.update {
-            it.copy(hasTronFrozenBalancesError = state is TronFrozenBalanceState.Error)
-        }
-    }
-
-    private fun updateTronFrozenBalanceDisplay(type: TronResourceType) {
-        val balances =
-            (tronFrozenBalances.value as? TronFrozenBalanceState.Loaded)?.balances ?: return
-        uiState.update {
-            it.copy(
-                tronBalanceAvailableOverride =
-                    balances.forResource(type).stripTrailingZeros().toPlainString()
-            )
-        }
-    }
-
-    private fun currentTronFrozenBalance(): BigDecimal? {
-        val type = uiState.value.tronResourceType ?: return null
-        val balances =
-            (tronFrozenBalances.value as? TronFrozenBalanceState.Loaded)?.balances ?: return null
-        return balances.forResource(type)
-    }
+    fun setTronResourceType(type: TronResourceType) = tronStakingService.setResourceType(type)
 
     private fun loadVaultName() {
         viewModelScope.launch {
@@ -913,7 +856,7 @@ constructor(
                         vaultId = vaultId,
                     )
 
-                    preSelectToken(
+                    tokenPreselectionService.preSelect(
                         preSelectedChainIds = preSelectedChainIds,
                         preSelectedTokenId = preSelectedTokenId,
                         forcePreselection = true,
@@ -931,7 +874,7 @@ constructor(
         if (!chainIdForAddition.isNullOrBlank() && chainIdNotInAccounts) {
             viewModelScope.launch {
                 addNativeTokenToVault(chainIdForAddition)
-                loadAccounts(vaultId)
+                accountsLoader.load(vaultId)
             }
         }
     }
@@ -1055,7 +998,7 @@ constructor(
         val newChainSelected = currentChain != newChain
         val isNotEvm = newChain.standard != TokenStandard.EVM
         if (newChainSelected && isNotEvm) {
-            preSelectToken(
+            tokenPreselectionService.preSelect(
                 preSelectedChainIds = listOf(newChain.id),
                 preSelectedTokenId = null,
                 forcePreselection = true,
@@ -1137,504 +1080,6 @@ constructor(
         uiState.update { it.copy(errorText = text) }
     }
 
-    private fun loadAccounts(vaultId: VaultId) {
-        loadAccountsJob?.cancel()
-        loadAccountsJob =
-            when (defiType) {
-                DeFiNavActions.WITHDRAW_RUJI ->
-                    viewModelScope.launch { loadRewardsAccount(vaultId) }
-
-                DeFiNavActions.WITHDRAW_USDC_CIRCLE ->
-                    viewModelScope.launch { loadCircleUSDCAccount(vaultId) }
-
-                null,
-                DeFiNavActions.BOND,
-                DeFiNavActions.STAKE_RUJI,
-                DeFiNavActions.STAKE_TCY,
-                DeFiNavActions.STAKE_STCY,
-                DeFiNavActions.UNSTAKE_STCY,
-                DeFiNavActions.MINT_YRUNE,
-                DeFiNavActions.MINT_YTCY,
-                DeFiNavActions.REDEEM_YRUNE,
-                DeFiNavActions.REDEEM_YTCY,
-                DeFiNavActions.DEPOSIT_USDC_CIRCLE,
-                DeFiNavActions.FREEZE_TRX ->
-                    viewModelScope.launch {
-                        accountsRepository
-                            .loadAddresses(vaultId)
-                            .map { addrs -> addrs.flatMap { it.accounts } }
-                            .collect(accounts)
-                    }
-
-                else ->
-                    viewModelScope.launch {
-                        accountsRepository
-                            .loadDeFiAddresses(vaultId, false)
-                            .map { addrs -> addrs.flatMap { it.accounts } }
-                            .collect(accounts)
-                    }
-            }
-    }
-
-    private suspend fun loadCircleUSDCAccount(vaultId: VaultId) {
-        val accountsLoaded =
-            accountsRepository.loadAddresses(vaultId).firstOrNull()?.flatMap { it.accounts }
-        val ethereumAccount =
-            accountsLoaded?.find { it.token.id.equals(Coins.Ethereum.ETH.id, true) }
-                ?: Account(
-                    token = Coins.Ethereum.ETH,
-                    tokenValue = TokenValue(BigInteger.ZERO, Coins.Ethereum.ETH),
-                    fiatValue = null,
-                    price = null,
-                )
-
-        val usdc = Coins.Ethereum.USDC.copy(address = ethereumAccount.token.address)
-
-        if (mscaAddress != null) {
-            val id = usdc.generateId(mscaAddress!!)
-            val cachedDetails = stakingDetailsRepository.getStakingDetailsById(vaultId, id)
-            val usdcCircleAccount =
-                Account(
-                    token = usdc,
-                    tokenValue =
-                        TokenValue(
-                            value = cachedDetails?.stakeAmount ?: BigInteger.ZERO,
-                            token = usdc,
-                        ),
-                    fiatValue = null,
-                    price = null,
-                )
-            accounts.value = listOf(ethereumAccount, usdcCircleAccount)
-        } else {
-            Timber.e("MSCA address not available for Circle USDC withdrawal")
-            accounts.value =
-                listOf(
-                    ethereumAccount,
-                    Account(
-                        token = usdc,
-                        tokenValue = TokenValue(value = BigInteger.ZERO, token = usdc),
-                        fiatValue = null,
-                        price = null,
-                    ),
-                )
-        }
-    }
-
-    private suspend fun loadRewardsAccount(vaultId: VaultId) {
-        val accountsLoaded =
-            accountsRepository.loadAddresses(vaultId).firstOrNull()?.flatMap { it.accounts }
-        val thorchainAccount =
-            accountsLoaded?.find { it.token.id.equals(Coins.ThorChain.RUNE.id, true) } ?: return
-
-        val rujiAccount =
-            accountsLoaded.find { it.token.id.equals(Coins.ThorChain.RUJI.id, true) } ?: return
-
-        val cachedDetails =
-            stakingDetailsRepository.getStakingDetailsByCoindId(vaultId, Coins.ThorChain.RUJI.id)
-
-        if (cachedDetails != null) {
-            val rewardsAccount =
-                Account(
-                    token = RUJI_REWARDS_COIN.copy(address = thorchainAccount.token.address),
-                    tokenValue =
-                        TokenValue(
-                            value = cachedDetails.rewards?.toBigInteger() ?: BigInteger.ZERO,
-                            token = RUJI_REWARDS_COIN,
-                        ),
-                    fiatValue = null,
-                    price = null,
-                )
-            accounts.value = listOf(rewardsAccount, thorchainAccount, rujiAccount)
-        } else {
-            accounts.value = emptyList()
-        }
-    }
-
-    private fun preSelectToken(
-        preSelectedChainIds: List<ChainId?>,
-        preSelectedTokenId: TokenId?,
-        forcePreselection: Boolean = false,
-    ) {
-        Timber.d("preSelectToken($preSelectedChainIds, $preSelectedTokenId, $forcePreselection)")
-
-        preSelectTokenJob?.cancel()
-        preSelectTokenJob =
-            viewModelScope.launch {
-                accounts.collect { accounts ->
-                    val preSelectedToken =
-                        if (defiType == null) {
-                            findPreselectedToken(accounts, preSelectedChainIds, preSelectedTokenId)
-                        } else {
-                            findDeFiPreselectedToken(
-                                accounts,
-                                preSelectedChainIds,
-                                preSelectedTokenId,
-                            )
-                        }
-
-                    Timber.d("Found a new token to pre select $preSelectedToken")
-
-                    // if user hasn't yet selected any token, preselect found token
-                    if (
-                        (forcePreselection || selectedTokenValue == null) &&
-                            preSelectedToken != null
-                    ) {
-                        selectToken(preSelectedToken)
-                    }
-                }
-            }
-    }
-
-    /**
-     * Returns first token found for tokenId or chainId or first token it all list, can return null
-     * if there's no tokens in the vault
-     */
-    private fun findPreselectedToken(
-        accounts: List<Account>,
-        preSelectedChainIds: List<ChainId?>,
-        preSelectedTokenId: TokenId?,
-    ): Coin? {
-        var searchByChainResult: Coin? = null
-
-        for (account in accounts) {
-            val accountToken = account.token
-            if (accountToken.id.equals(preSelectedTokenId, ignoreCase = true)) {
-                // if we find token by id, return it asap
-                return accountToken
-            }
-            if (
-                searchByChainResult == null &&
-                    preSelectedChainIds.contains(accountToken.chain.id) &&
-                    accountToken.isNativeToken
-            ) {
-                // if we find token by chain, remember it and return later if nothing else found
-                searchByChainResult = accountToken
-            }
-        }
-
-        // if user selected none, or nothing was found, select the first token
-        return searchByChainResult ?: accounts.firstOrNull()?.token
-    }
-
-    private fun findDeFiPreselectedToken(
-        accounts: List<Account>,
-        preSelectedChainIds: List<ChainId?>,
-        preSelectedTokenId: TokenId?,
-    ): Coin? {
-        for (account in accounts) {
-            val accountToken = account.token
-            if (accountToken.id.equals(preSelectedTokenId, ignoreCase = true)) {
-                return accountToken
-            }
-        }
-
-        // default coins, in case the account does not exist
-        val defaultCoin =
-            when (defiType) {
-                DeFiNavActions.STAKE_RUJI,
-                DeFiNavActions.UNSTAKE_RUJI -> Coins.ThorChain.RUJI
-
-                DeFiNavActions.STAKE_TCY,
-                DeFiNavActions.UNSTAKE_TCY -> Coins.ThorChain.TCY
-
-                DeFiNavActions.MINT_YRUNE -> Coins.ThorChain.RUNE
-                DeFiNavActions.MINT_YTCY -> Coins.ThorChain.TCY
-                DeFiNavActions.BOND -> Coins.ThorChain.RUNE
-                DeFiNavActions.UNBOND -> Coins.ThorChain.RUNE
-                DeFiNavActions.WITHDRAW_RUJI -> RUJI_REWARDS_COIN
-                DeFiNavActions.REDEEM_YRUNE -> Coins.ThorChain.yRUNE
-                DeFiNavActions.REDEEM_YTCY -> Coins.ThorChain.yTCY
-                DeFiNavActions.DEPOSIT_USDC_CIRCLE -> Coins.Ethereum.USDC
-                DeFiNavActions.WITHDRAW_USDC_CIRCLE -> Coins.Ethereum.USDC
-                DeFiNavActions.STAKE_STCY -> Coins.ThorChain.TCY
-                DeFiNavActions.UNSTAKE_STCY -> Coins.ThorChain.sTCY
-                DeFiNavActions.STAKE_CACAO,
-                DeFiNavActions.UNSTAKE_CACAO,
-                DeFiNavActions.ADD_LP,
-                DeFiNavActions.REMOVE_LP -> Coins.MayaChain.CACAO
-                DeFiNavActions.FREEZE_TRX,
-                DeFiNavActions.UNFREEZE_TRX -> Coins.Tron.TRX
-                null -> findPreselectedToken(accounts, preSelectedChainIds, preSelectedTokenId)
-            }
-
-        return defaultCoin
-    }
-
-    private fun calculateGasTokenBalance() {
-        viewModelScope.launch {
-            selectedToken
-                .filterNotNull()
-                .map {
-                    if (it.isNativeToken) {
-                        null
-                    } else {
-                        accounts.value
-                            .find { account ->
-                                account.token.isNativeToken && account.token.chain == it.chain
-                            }
-                            ?.tokenValue
-                    }
-                }
-                .collect { gasTokenBalance ->
-                    if (gasTokenBalance == null) {
-                        uiState.update { it.copy(gasTokenBalance = null) }
-                    } else {
-                        uiState.update {
-                            it.copy(
-                                gasTokenBalance =
-                                    UiText.DynamicString(mapTokenValueToString(gasTokenBalance))
-                            )
-                        }
-                    }
-                }
-        }
-    }
-
-    @OptIn(FlowPreview::class)
-    private fun calculateGasFees() {
-        viewModelScope.launch {
-            combine(
-                    combine(
-                            selectedToken.filterNotNull(),
-                            addressFieldState.textAsFlow(),
-                            memoFieldState.textAsFlow(),
-                            tokenAmountFieldState.textAsFlow(),
-                            recalculateGasFee,
-                        ) { token, dst, memo, tokenAmount, nonce ->
-                            GasFeeInput(
-                                token,
-                                dst.asAddressInput(),
-                                memo.toString(),
-                                tokenAmount,
-                                nonce,
-                            )
-                        }
-                        .debounce(350)
-                        .distinctUntilChanged()
-                        .mapNotNull { (token, dst, memo, tokenAmount) ->
-                            val vault = vault ?: return@mapNotNull null
-                            val tokenAmount = tokenAmount.toString().toBigDecimalOrNull()
-
-                            val tokenAmountInt =
-                                tokenAmount?.movePointRight(token.decimal)?.toBigInteger()
-                                    ?: return@mapNotNull null
-
-                            val chain = token.chain
-                            val blockchainTransaction =
-                                Transfer(
-                                    coin = token,
-                                    vault =
-                                        VaultData(
-                                            vaultHexChainCode = vault.hexChainCode,
-                                            vaultHexPublicKey = vault.getPubKeyByChain(chain),
-                                        ),
-                                    amount = tokenAmountInt,
-                                    to = addressManager.resolvedDstAddress.value ?: dst,
-                                    memo = memo,
-                                    isMax = false,
-                                )
-
-                            val fees =
-                                withContext(Dispatchers.IO) {
-                                    feeServiceComposite.calculateFees(blockchainTransaction)
-                                }
-                            val nativeCoin =
-                                withContext(Dispatchers.IO) {
-                                    tokenRepository.getNativeToken(chain.id)
-                                }
-
-                            TokenValue(value = fees.amount, token = nativeCoin)
-                        }
-                        .catch { Timber.e(it) },
-                    gasSettings,
-                    specific,
-                ) { gasFee, gasSettings, specific ->
-                    this@SendFormViewModel.gasFee.value =
-                        adjustGasFee(gasFee, gasSettings, specific)
-                }
-                .collect()
-        }
-    }
-
-    private fun collectPlanFee() {
-        viewModelScope.launch {
-            combine(
-                    selectedToken.filterNotNull(),
-                    addressFieldState.textAsFlow(),
-                    tokenAmountFieldState.textAsFlow(),
-                    specific.filterNotNull(),
-                    memoFieldState.textAsFlow(),
-                ) { token, dstAddress, tokenAmount, specific, memo ->
-                    PlanFeeInput(token, dstAddress, tokenAmount, specific, memo)
-                }
-                .combine(recalculateGasFee) { data, _ -> data }
-                .mapNotNull { (token, dstAddress, tokenAmount, specific, memo) ->
-                    try {
-                        val chain = token.chain
-                        if (chain.standard != TokenStandard.UTXO || chain == Chain.Cardano) {
-                            planFee.value = 1
-                            return@mapNotNull null
-                        }
-
-                        val vaultId =
-                            vaultId
-                                ?: throw InvalidTransactionDataException(
-                                    UiText.StringResource(string.send_error_no_token)
-                                )
-
-                        val resolvedDstAddress =
-                            addressParserRepository.resolveName(dstAddress.asAddressInput(), chain)
-                        val tokenAmountInt =
-                            tokenAmount
-                                .toString()
-                                .toBigDecimal()
-                                .movePointRight(token.decimal)
-                                .toBigInteger()
-
-                        val plan =
-                            bitcoinPlanService.getPlan(
-                                vaultId,
-                                token,
-                                resolvedDstAddress,
-                                tokenAmountInt,
-                                specific,
-                                memo.toString(),
-                            )
-
-                        planFee.value = plan.fee
-                        planBtc.value = plan
-                    } catch (e: Exception) {
-                        Timber.e(e)
-                    }
-                }
-                .collect()
-        }
-    }
-
-    private fun collectMaxAmount() {
-        viewModelScope.launch {
-            amountManager.isMaxAmount.collect { isMax ->
-                val chain = selectedAccount?.token?.chain ?: return@collect
-                // Only require to re-trigger utxo chains, due to no change output utxo and
-                // therefore
-                // less fees
-                if (chain.standard == TokenStandard.UTXO && chain != Chain.Cardano) {
-                    val spec =
-                        specific.value?.blockChainSpecific as? BlockChainSpecific.UTXO
-                            ?: return@collect
-                    val updatedSpec =
-                        specific.value?.copy(blockChainSpecific = spec.copy(sendMaxAmount = isMax))
-                    specific.value = updatedSpec
-                }
-            }
-        }
-    }
-
-    private fun collectEstimatedFee() {
-        viewModelScope.launch {
-            combine(
-                    selectedToken.filterNotNull(),
-                    gasFee.filterNotNull(),
-                    gasSettings,
-                    planFee.filterNotNull(),
-                ) { token, gasFee, gasSettings, planFee ->
-                    val chain = token.chain
-                    val evmGasSettings = gasSettings as? GasSettings.Eth
-                    val estimatedFee =
-                        gasFeeToEstimatedFee(
-                            GasFeeParams(
-                                gasLimit =
-                                    if (evmGasSettings != null) evmGasSettings.gasLimit
-                                    else BigInteger.valueOf(1),
-                                gasFee =
-                                    selectGasFeeForFeeEstimation(
-                                        chain = chain,
-                                        gasFee = gasFee,
-                                        planFee = planFee,
-                                        evmGasSettings = evmGasSettings,
-                                    ),
-                                selectedToken = token,
-                                perUnit = true,
-                            )
-                        )
-
-                    uiState.update {
-                        it.copy(
-                            estimatedFee = UiText.DynamicString(estimatedFee.formattedFiatValue),
-                            totalGas = UiText.DynamicString(estimatedFee.formattedTokenValue),
-                        )
-                    }
-                }
-                .collect()
-        }
-    }
-
-    @OptIn(FlowPreview::class)
-    private fun calculateSpecific() {
-        viewModelScope.launch {
-            // dstAddress is forwarded to getSpecific() for every chain so that
-            // TRON fee estimation can account for bandwidth delegated to the receiver.
-            // Recomputing specifics on every keystroke would be wasteful, so the
-            // address is debounced before triggering the recalculation.
-            val dstAddressFlow =
-                addressFieldState
-                    .textAsFlow()
-                    .map { it.toString().asAddressInput() }
-                    .debounce(300)
-                    .distinctUntilChanged()
-
-            combine(selectedToken.filterNotNull(), gasFee.filterNotNull(), dstAddressFlow) {
-                    token,
-                    gasFee,
-                    dstAddress ->
-                    val chain = token.chain
-                    val srcAddress = token.address
-                    advanceGasUiRepository.updateTokenStandard(token.chain.standard)
-
-                    val validDstAddress =
-                        dstAddress.takeIf {
-                            it.isNotBlank() && chainAccountAddressRepository.isValid(chain, it)
-                        }
-
-                    try {
-                        val spec =
-                            blockChainSpecificRepository.getSpecific(
-                                chain,
-                                srcAddress,
-                                token,
-                                gasFee,
-                                isSwap = false,
-                                isMaxAmountEnabled = false,
-                                isDeposit = false,
-                                dstAddress = validDstAddress,
-                            )
-                        specific.value = spec
-                        advanceGasUiRepository.updateBlockChainSpecific(spec.blockChainSpecific)
-                        advanceGasUiRepository.showIcon()
-                        uiState.update { it.copy(specific = spec) }
-                    } catch (e: Exception) {
-                        Timber.e(e)
-                    }
-                }
-                .collect()
-        }
-    }
-
-    private fun adjustGasFee(
-        gasFee: TokenValue,
-        gasSettings: GasSettings?,
-        spec: BlockChainSpecificAndUtxo?,
-    ) =
-        gasFee.copy(
-            value =
-                if (
-                    gasSettings is GasSettings.UTXO &&
-                        spec?.blockChainSpecific is BlockChainSpecific.UTXO
-                ) {
-                    gasSettings.byteFee
-                } else gasFee.value
-        )
-
     private fun loadSelectedCurrency() {
         viewModelScope.launch {
             appCurrency.collect { appCurrency ->
@@ -1702,32 +1147,5 @@ constructor(
         uiState.update { it.copy(expandedSection = section) }
     }
 
-    fun refreshGasFee() {
-        viewModelScope.launch {
-            uiState.update { it.copy(isRefreshing = true) }
-            recalculateGasFee.update { it + 1 }
-            // Rapid toggling of isRefreshing can cause the initial true value to be skipped,
-            // displaying only the false value in the UI resulting in the swipe refresh being
-            // frozen.
-            // this line prevent missing true value in these cases.
-            delay(100)
-            uiState.update { it.copy(isRefreshing = false) }
-        }
-    }
+    fun refreshGasFee() = gasFeeOrchestrator.refresh()
 }
-
-private data class GasFeeInput(
-    val token: Coin,
-    val dst: String,
-    val memo: String,
-    val tokenAmount: CharSequence,
-    val nonce: Long = 0,
-)
-
-private data class PlanFeeInput(
-    val token: Coin,
-    val dstAddress: CharSequence,
-    val tokenAmount: CharSequence,
-    val specific: BlockChainSpecificAndUtxo,
-    val memo: CharSequence,
-)
